backend/requirements.txt
Babel==2.11.0
dj-rest-auth==2.2.6
Django==4.0.4
django-allauth==0.51.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==22.1
django-mptt==0.14.0
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django2-babel==0.6.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
phonenumbers==8.13.3
Pillow==9.1.1
PyJWT==2.4.0
python-dotenv==0.21.0